+++ superset/tasks/celery_app.py
@@ -32,7 +32,8 @@ flask_app = create_app()
 
 # Need to import late, as the celery_app will have been setup by "create_app()"
 # pylint: disable=wrong-import-position, unused-import
-from . import cache, schedules, scheduler  # isort:skip
+from . import schedules, scheduler  # isort:skip
+import caching.cache # isort:skip
 
 # Export the celery app globally for Celery (as run on the cmd line) to find
 app = celery_app
