+++ tests/integration_tests/strategy_tests.py
@@ -23,27 +23,19 @@ from tests.integration_tests.fixtures.birth_names_dashboard import (
     load_birth_names_dashboard_with_slices,
 )
 
-from sqlalchemy import String, Date, Float
-
 import pytest
-import pandas as pd
-
-from superset.models.slice import Slice
-from superset.utils.core import get_example_database
 
 from superset import db
 
 from superset.models.core import Log
 from superset.models.tags import get_tag, ObjectTypes, TaggedObject, TagTypes
-from superset.tasks.cache import (
+from superset.tasks.caching.cache_strategy import (
     DashboardTagsStrategy,
     get_form_data,
     TopNDashboardsStrategy,
 )
 
 from .base_tests import SupersetTestCase
-from .dashboard_utils import create_dashboard, create_slice, create_table_for_dashboard
-from .fixtures.unicode_dashboard import load_unicode_dashboard_with_slice
 
 URL_PREFIX = "http://0.0.0.0:8081"
